Rework of matrix implementation
New underlying data type for matrix implementation. It now uses vector
instead of array. Had some trouble with the macro solution when I
started to do work with inverting matrices. Hopefully it will be easier
with this implementation. However this implementation is not that safe
since it will crash at runtime when doing illegal operations, such as
multiplying a 4x4 matrix with a 3x3 matrix.